[2604-CHORE-139] extract EventForm, Pill, AdminCalendarClient from calendar page
